Move common mailer behaviour to common concern
